requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.56.0
Pillow>=9.0.0
numpy>=1.21.0
tqdm>=4.60.0
pyyaml>=6.0
